requirements.txt
annotated-types==0.7.0
astroid==3.3.10
asttokens==3.0.0
beartype==0.20.2
black==25.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
cmcrameri==1.9
cmocean==4.0.3
colorcet==3.1.0
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
dill==0.4.0
docker-pycreds==0.4.0
ecos==2.0.14
einops==0.8.1
einx==0.3.0
executing==2.2.0
fastremap==1.17.6
filelock==3.18.0
fill_voids==2.1.1
fonttools==4.58.0
frozendict==2.4.6
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
grad-cam==1.5.5
h5py==3.14.0
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
ipdb==0.13.13
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
kiwisolver==1.4.8
lazy_loader==0.4
lightning-utilities==0.14.3
lungmask==0.2.20
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.4.2
numexpr==2.10.2
numpy==2.2.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
osqp==0.6.7.post3
packaging==25.0
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
pooch==1.8.2
prompt_toolkit==3.0.51
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.4
pydantic_core==2.33.2
pydicom==3.0.1
Pygments==2.19.1
pylint==3.3.7
pyparsing==3.2.3
pytest==8.3.5
pytest-sugar==1.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyvista==0.46.3
PyYAML==6.0.2
qdldl==0.1.7.post5
requests==2.32.3
rich==14.0.0
scikit-image==0.25.2
scikit-learn==1.6.1
scikit-survival==0.24.1
scipy==1.15.3
scooby==0.10.1
sentry-sdk==2.28.0
setproctitle==1.3.6
setuptools==80.4.0
simpleitk==2.5.2
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.14.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.9.30
tomlkit==0.13.2
torch==2.7.0
torchaudio==2.7.0
torchmetrics==1.7.1
torchvision==0.22.0
tqdm==4.67.1
traitlets==5.14.3
triton==3.3.0
ttach==0.0.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vector-quantize-pytorch==1.22.16
vtk==9.5.1
wandb==0.19.11
wcwidth==0.2.13
wheel==0.45.1
